requirements.txt
fastapi
uvicorn
pydantic
PyYAML
requests
pytest
